pyproject.toml
[project]
name = "visascraper"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-rucaptcha (>=6.4.0,<7.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=5.4.0,<6.0.0)",
    "curl-cffi (>=0.11.2,<0.12.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)",
    "google-api-python-client (>=2.171.0,<3.0.0)",
    "dishka (>=1.6.0,<2.0.0)",
    "pydrive2 (>=1.21.3,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "tls-client (>=1.0.1,<2.0.0)",
    "mega (>=0.2.11,<0.3.0)",
    "yadisk (>=3.4.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "visascraper", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run = "visascraper.main:main"
